fix: race condition in resolver when the data source is cancelled (#671)
If the source get cancelled while resolver had just checked the Context
and started to receive the next payload from the data source we would
get stuck. Instead of using a default in the select, select between the
Done() channel and the channel receiving payload. This way any
cancellation
will immediately stop blocking trying to receive the next payload.